Update WhileLoop.java
Program that exercutes a while loop through a JOptionPane.
It contains a state that can be changed by the User during run time. This is mainly through prompting the user if they want to continue or not, testing the condition while its true.

It has 2 options that the user can choose from which are "1"  and "2"
case 0 being Yes with a switch statement that continues to exercuute the program and case 1 being No with a switch statement that outputs "User, Terminate the Program